clip_automation_poc/pyproject.toml
[project]
name = "clip-automation-poc"
version = "0.1.0"
description = "Urban Dynamics - Real Estate Data Visualization PoC"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "jupyterlab>=4.5.2",
    "pandas>=2.0.0,<3.0.0",
    "numpy>=2.0.0",
    "pydeck>=0.9.0",
    "pyarrow>=18.0.0",
    "python-dotenv>=1.2.1",
    "polars>=1.37.1",
    "geopandas>=1.0.0",
    "shapely>=2.0.0",
    "scipy>=1.14.0",
    "prophet>=1.1.5",
    "lightgbm>=4.5.0",
    "scikit-learn>=1.5.0",
    "plotly>=6.5.2",
    "streamlit>=1.40.0",
]


[dependency-groups]
dev = [
    "ruff>=0.14.10",
]

[tool.ruff]
# 行の長さ制限
line-length = 120

# Pythonのターゲットバージョン
target-version = "py313"

# 有効にするルール
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
]

# 無視するルール
ignore = [
    "E501",  # line too long (handled by line-length)
]

# 除外するファイル/ディレクトリ
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

[tool.ruff.format]
# クォートスタイル
quote-style = "double"

# インデントスタイル
indent-style = "space"

# 行末での文字列の自動結合を行う
skip-magic-trailing-comma = false

[tool.ruff.isort]
# importの順序
force-single-line = false
lines-between-types = 1


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/real_state_geo_core"]
